requirements.txt
absl-py==2.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
beautifulsoup4==4.13.4
boto3==1.39.7
botocore==1.39.7
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
faiss-cpu==1.11.0.post1
feedparser==6.0.11
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
minio==7.2.15
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==5.29.5
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2

